pyproject.toml
[tool.poetry]
name = "utility-scripts"
version = "0.1.0"
description = "Utility scripts for working with Gemini API and other tools"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
torch = "^2.3.0"
transformers = "^4.43.3"
diffusers = "^0.30.0"
accelerate = "^0.34.0"
safetensors = "^0.4.5"
tqdm = "^4.66.4"

[tool.poetry.scripts]
generate-car-parts = "utility_scripts.generate_car_parts_gemini_api.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
